Rename FloatingActionButtonItem and FloatingActionButtonBase
