+++ js/ui/speed_dial_action.js
@@ -4,4 +4,4 @@
  * @module ui/speed_dial_action
  * @export default
  */
-module.exports = require("./floating_action_button/speed_dial_action");
+module.exports = require("./speed_dial_action/speed_dial_action");
